rust cli: consolidate integration tests into a single tests/cli.rs
Fold the completion tests into cli.rs and remove the separate
completion.rs. A single file keeps all e2e tests in one test binary with
shared helpers (no tests/common/ needed) and one readable ordering:
exit codes, stdin pipes, completion, surface smoke.
